Resolve regional URL in the acceptance test client
The acceptance test helper built its client from the unresolved host, so
every request made by acctests.providerClient was redirected by the region
service before reaching the API. Since #885 the provider resolves the
regional URL in Configure, but test.TwingateClient constructs its client
directly and never inherited that change.

Extract the resolution into twingate/internal/client as ResolveRegionalURL
and call it from both Configure and test.TwingateClient, so there is one
implementation rather than two duplicated URL expressions. The test helper
client is a package-level singleton, so this is one resolve per test binary
instead of one redirect per request.

The moved helper now builds its request with http.NewRequestWithContext and
Do rather than Get: the noctx linter rejects the bare Get in the client
package, and threading the context gives the resolve cancellation.

Fixes #945
